requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
google-generativeai
numpy
faiss-cpu
python-dotenv
